+++ config/config.exs
@@ -340,6 +340,7 @@ config :pleroma, :activitypub,
   unfollow_blocked: true,
   outgoing_blocks: true,
   follow_handshake_timeout: 500,
+  note_replies_output_limit: 5,
   sign_object_fetches: true
 
 config :pleroma, :streamer,
@@ -623,10 +624,6 @@ config :pleroma, :web_cache_ttl,
 config :pleroma, :modules, runtime_dir: "instance/modules"
 
 config :pleroma, configurable_from_database: false
-
-config :pleroma, :mastodon_compatibility,
-  # https://git.pleroma.social/pleroma/pleroma/issues/1505
-  federated_note_replies_limit: 5
 
 config :swarm, node_blacklist: [~r/myhtml_.*$/]
 # Import environment specific config. This must remain at the bottom